requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1